+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # Gpdviz
 
+2017-10-24
+
+- add option to import existing models in json files
+
 2017-10-23
 
 - first slick-based version